+++ src/iqm/qiskit_iqm/fake_backends/fake_apollo.py
@@ -11,11 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Fake (i.e. simulated) backend for IQM Apollo QPU.
+"""Fake (i.e. simulated) backend for IQM's 20-qubit Apollo architecture
 """
-from iqm_client import QuantumArchitectureSpecification
+from iqm.iqm_client import QuantumArchitectureSpecification
 
-from qiskit_iqm.fake_backends.iqm_fake_backend import IQMErrorProfile, IQMFakeBackend
+from iqm.qiskit_iqm.fake_backends.iqm_fake_backend import IQMErrorProfile, IQMFakeBackend
 
 
 def IQMFakeApollo() -> IQMFakeBackend: