+++ tests/fake_backends/test_iqm_fake_backend.py
@@ -18,7 +18,7 @@ from qiskit import QuantumCircuit
 from qiskit.providers import JobV1
 from qiskit_aer.noise.noise_model import NoiseModel
 
-from qiskit_iqm import IQMFakeBackend
+from iqm.qiskit_iqm import IQMFakeBackend
 
 
 @pytest.fixture
@@ -227,3 +227,6 @@ def test_noise_model_contains_all_errors(backend):
     Test that the noise model contains all necessary errors.
     """
     assert set(backend.noise_model.noise_instructions) == {"r", "cz", "measure"}
+
+    # Assert that CZ gate error is applied independent of argument order in gate specification
+    assert set(backend.noise_model._local_quantum_errors["cz"].keys()) == set([(0, 1), (1, 0), (1, 2), (2, 1)])