+++ tests/test_iqm_backend.py
@@ -21,10 +21,10 @@ from qiskit import QuantumCircuit
 from qiskit.compiler import transpile
 from qiskit.providers import Options
 
-from qiskit_iqm.iqm_backend import IQMBackendBase
+from iqm.qiskit_iqm.iqm_backend import IQMBackendBase
 
 
-class TestIQMBackend(IQMBackendBase):
+class DummyIQMBackend(IQMBackendBase):
     """Dummy implementation for abstract methods of IQMBacked, so that instances can be created
     and the rest of functionality tested."""
 
@@ -42,11 +42,11 @@ class TestIQMBackend(IQMBackendBase):
 
 @pytest.fixture
 def backend(linear_architecture_3q):
-    return TestIQMBackend(linear_architecture_3q)
+    return DummyIQMBackend(linear_architecture_3q)
 
 
 def test_qubit_name_to_index_to_qubit_name(adonis_architecture_shuffled_names):
-    backend = TestIQMBackend(adonis_architecture_shuffled_names)
+    backend = DummyIQMBackend(adonis_architecture_shuffled_names)
 
     correct_idx_name_associations = set(enumerate(['QB1', 'QB2', 'QB3', 'QB4', 'QB5']))
     assert all(backend.index_to_qubit_name(idx) == name for idx, name in correct_idx_name_associations)
@@ -76,7 +76,7 @@ def test_transpile(backend):
 def test_validate_compatible_architecture(
     adonis_architecture, adonis_architecture_shuffled_names, linear_architecture_3q
 ):
-    backend = TestIQMBackend(adonis_architecture)
+    backend = DummyIQMBackend(adonis_architecture)
     assert backend.validate_compatible_architecture(adonis_architecture) is True
     assert backend.validate_compatible_architecture(adonis_architecture_shuffled_names) is True
     assert backend.validate_compatible_architecture(linear_architecture_3q) is False