+++ tests/test_iqm_provider.py
@@ -14,21 +14,21 @@
 
 """Testing IQM provider.
 """
+from collections.abc import Sequence
 from importlib.metadata import version
-from numbers import Number
 import uuid
 
-from iqm_client import HeraldingMode, IQMClient, RunResult, RunStatus
 from mockito import ANY, mock, patch, when
 import numpy as np
 import pytest
 from qiskit import QuantumCircuit, execute
-from qiskit.circuit import Parameter, ParameterExpression
-from qiskit.circuit.library import RGate
+from qiskit.circuit import Parameter
+from qiskit.circuit.library import RGate, RXGate, RYGate, XGate, YGate
 from qiskit.compiler import transpile
 
-from qiskit_iqm import IQMBackend, IQMJob, IQMProvider
-from qiskit_iqm.iqm_provider import IQMFacadeBackend
+from iqm.iqm_client import HeraldingMode, IQMClient, RunResult, RunStatus
+from iqm.qiskit_iqm import IQMBackend, IQMJob, IQMProvider
+from iqm.qiskit_iqm.iqm_provider import IQMFacadeBackend
 
 
 @pytest.fixture
@@ -69,6 +69,9 @@ def job_id():
 def test_default_options(backend):
     assert backend.options.shots == 1024
     assert backend.options.calibration_set_id is None
+    assert backend.options.circuit_duration_check is True
+    assert backend.options.heralding_mode == HeraldingMode.NONE
+    assert backend.options.circuit_callback is None
 
 
 def test_retrieve_job(backend):
@@ -77,8 +80,18 @@ def test_retrieve_job(backend):
     assert job.job_id() == 'a job id'
 
 
-def test_max_circuits(backend):
+def test_default_max_circuits(backend):
     assert backend.max_circuits is None
+
+
+def test_set_max_circuits(backend):
+    assert backend.max_circuits is None
+
+    backend.max_circuits = 17
+    assert backend.max_circuits == 17
+
+    backend.max_circuits = 168
+    assert backend.max_circuits == 168
 
 
 def test_qubit_name_to_index_to_qubit_name(adonis_architecture_shuffled_names):
@@ -104,6 +117,14 @@ def test_serialize_circuit_raises_error_for_unsupported_instruction(backend, cir
     circuit.sx(0)
     with pytest.raises(ValueError, match=f"Instruction 'sx' in the circuit '{circuit.name}' is not natively supported"):
         backend.serialize_circuit(circuit)
+
+
+def test_serialize_circuit_does_not_raise_for_x_rx_y_ry(backend, circuit):
+    circuit.x(0)
+    circuit.rx(0.123, 0)
+    circuit.y(0)
+    circuit.ry(0.321, 0)
+    backend.serialize_circuit(circuit)
 
 
 def test_serialize_circuit_raises_error_for_unsupported_metadata(backend, circuit):
@@ -135,23 +156,26 @@ def test_serialize_circuit_maps_r_gate(circuit, gate, expected_angle, expected_p
     assert instr.args['phase_t'] == expected_phase
 
 
-def test_serialize_handles_parameter_expressions(circuit, backend):
-    theta = Parameter('θ')
-    phi = Parameter('φ')
-    circuit.r(theta, phi, 0)
-    circuit_bound = circuit.bind_parameters({theta: np.pi, phi: 0})
-
-    # First make sure that circuit_bound does indeed represent parameters as ParameterExpression
-    assert len(circuit_bound.data) == 1
-    instruction = circuit_bound.data[0][0]
-    assert all(isinstance(param, ParameterExpression) for param in instruction.params)
-
-    # Now check that serialization correctly handles ParameterExpression
-    circuit_ser = backend.serialize_circuit(circuit_bound)
+@pytest.mark.parametrize(
+    'gate, expected_angle, expected_phase',
+    [
+        (XGate(), 1 / 2, 0),
+        (RXGate(theta=np.pi / 2), 1 / 4, 0),
+        (RXGate(theta=2 * np.pi / 3), 1 / 3, 0),
+        (YGate(), 1 / 2, 1 / 4),
+        (RYGate(theta=np.pi / 2), 1 / 4, 1 / 4),
+        (RYGate(theta=2 * np.pi / 3), 1 / 3, 1 / 4),
+    ],
+)
+def test_serialize_circuit_maps_x_rx_y_ry_gates(backend, circuit, gate, expected_angle, expected_phase):
+    circuit.append(gate, [0])
+    circuit_ser = backend.serialize_circuit(circuit)
     assert len(circuit_ser.instructions) == 1
-    iqm_instruction = circuit_ser.instructions[0]
-    assert isinstance(iqm_instruction.args['angle_t'], Number)
-    assert isinstance(iqm_instruction.args['phase_t'], Number)
+    instr = circuit_ser.instructions[0]
+    assert instr.name == 'phased_rx'
+    assert instr.qubits == ('0',)
+    assert instr.args['angle_t'] == expected_angle
+    assert instr.args['phase_t'] == expected_phase
 
 
 def test_serialize_circuit_maps_cz_gate(circuit, backend):
@@ -280,6 +304,14 @@ def test_run_sets_circuit_metadata_to_the_job(backend):
     assert job.circuit_metadata == [circuit_1.metadata, circuit_2.metadata]
 
 
+@pytest.mark.parametrize('shots', [13, 978, 1137])
+def test_run_with_custom_number_of_shots(backend, circuit, submit_circuits_default_kwargs, job_id, shots):
+    circuit.measure(0, 0)
+    kwargs = submit_circuits_default_kwargs | {'shots': shots, 'qubit_mapping': {'0': 'QB1'}}
+    when(backend.client).submit_circuits(ANY, **kwargs).thenReturn(job_id)
+    backend.run(circuit, shots=shots)
+
+
 @pytest.mark.parametrize(
     'calibration_set_id', ['67e77465-d90e-4839-986e-9270f952b743', uuid.UUID('67e77465-d90e-4839-986e-9270f952b743')]
 )
@@ -322,6 +354,37 @@ def test_run_with_heralding_mode_zeros(backend, circuit, submit_circuits_default
     backend.run([circuit], heralding_mode='zeros')
 
 
+# mypy: disable-error-code="attr-defined"
+def test_run_with_circuit_callback(backend, job_id, submit_circuits_default_kwargs):
+    qc1 = QuantumCircuit(3)
+    qc1.measure_all()
+    qc2 = QuantumCircuit(3)
+    qc2.r(np.pi, 0.3, 0)
+    qc2.measure_all()
+
+    def sample_callback(circuits) -> None:
+        assert isinstance(circuits, Sequence)
+        assert all(isinstance(c, QuantumCircuit) for c in circuits)
+        assert len(circuits) == 2
+        assert circuits[0].name == qc1.name
+        assert circuits[1].name == qc2.name
+        sample_callback.called = True
+
+    sample_callback.called = False
+
+    kwargs = submit_circuits_default_kwargs | {'qubit_mapping': {'0': 'QB1', '1': 'QB2', '2': 'QB3'}}
+    when(backend.client).submit_circuits(ANY, **kwargs).thenReturn(job_id)
+    backend.run([qc1, qc2], circuit_callback=sample_callback)
+    assert sample_callback.called is True
+
+
+def test_run_with_unknown_option(backend, circuit, job_id):
+    circuit.measure_all()
+    when(backend.client).submit_circuits(...).thenReturn(job_id)
+    with pytest.warns(Warning, match=r'Unknown backend option\(s\)'):
+        backend.run(circuit, to_option_or_not_to_option=17)
+
+
 def test_run_batch_of_circuits(backend, circuit, submit_circuits_default_kwargs, job_id):
     theta = Parameter('theta')
     theta_range = np.linspace(0, 2 * np.pi, 3)
@@ -361,7 +424,7 @@ def test_get_backend(linear_architecture_3q):
     assert isinstance(backend, IQMBackend)
     assert backend.client._base_url == url
     assert backend.num_qubits == 3
-    assert set(backend.coupling_map.get_edges()) == {(0, 1), (1, 2)}
+    assert set(backend.coupling_map.get_edges()) == {(0, 1), (1, 0), (1, 2), (2, 1)}
 
 
 def test_client_signature():
@@ -371,7 +434,7 @@ def test_client_signature():
     assert f'qiskit-iqm {version("qiskit-iqm")}' in backend.client._signature
 
 
-def test_get_facade_backend(adonis_architecture):
+def test_get_facade_backend(adonis_architecture, adonis_coupling_map):
     url = 'http://some_url'
     when(IQMClient).get_quantum_architecture().thenReturn(adonis_architecture)
 
@@ -381,7 +444,7 @@ def test_get_facade_backend(adonis_architecture):
     assert isinstance(backend, IQMFacadeBackend)
     assert backend.client._base_url == url
     assert backend.num_qubits == 5
-    assert set(backend.coupling_map.get_edges()) == {(0, 2), (1, 2), (3, 2), (4, 2)}
+    assert set(backend.coupling_map.get_edges()) == adonis_coupling_map
 
 
 def test_get_facade_backend_raises_error_non_matching_architecture(linear_architecture_3q):